+++ spring-cloud-netflix-eureka-client/src/main/java/org/springframework/cloud/netflix/ribbon/eureka/RibbonEurekaAutoConfiguration.java
@@ -17,18 +17,10 @@
 package org.springframework.cloud.netflix.ribbon.eureka;
 
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
-import org.springframework.boot.autoconfigure.condition.AllNestedConditions;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.cloud.netflix.ribbon.RibbonAutoConfiguration;
 import org.springframework.cloud.netflix.ribbon.RibbonClients;
-import org.springframework.cloud.netflix.ribbon.SpringClientFactory;
 import org.springframework.context.annotation.Configuration;
-
-import com.netflix.discovery.EurekaClient;
-import com.netflix.niws.loadbalancer.DiscoveryEnabledNIWSServerList;
 
 /**
  * Spring configuration for configuring Ribbon defaults to be Eureka based 
@@ -44,21 +36,5 @@ import com.netflix.niws.loadbalancer.DiscoveryEnabledNIWSServerList;
 @RibbonClients(defaultConfiguration = EurekaRibbonClientConfiguration.class)
 public class RibbonEurekaAutoConfiguration {
 
-	static class OnRibbonAndEurekaEnabledCondition extends AllNestedConditions {
 
-		public OnRibbonAndEurekaEnabledCondition() {
-			super(ConfigurationPhase.REGISTER_BEAN);
-		}
-
-		@ConditionalOnClass(DiscoveryEnabledNIWSServerList.class)
-		@ConditionalOnBean(SpringClientFactory.class)
-		@ConditionalOnProperty(value = "ribbon.eureka.enabled", matchIfMissing = true)
-		static class Defaults {}
-		
-		@ConditionalOnBean(EurekaClient.class)
-		static class EurekaBeans {}
-
-		@ConditionalOnProperty(value = "eureka.client.enabled", matchIfMissing = true)
-		static class OnEurekaClientEnabled {}
-	}
 }
