+++ spring-cloud-netflix-eureka-client/src/test/java/org/springframework/cloud/netflix/eureka/ConditionalOnRefreshScopeTests.java
@@ -20,15 +20,14 @@ import org.junit.After;
 import org.junit.Test;
 import org.springframework.boot.Banner.Mode;
 import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.test.context.assertj.AssertableApplicationContext;
 import org.springframework.cloud.autoconfigure.RefreshAutoConfiguration;
 import org.springframework.cloud.netflix.eureka.EurekaClientAutoConfiguration.ConditionalOnRefreshScope;
-import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNotNull;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.springframework.boot.WebApplicationType.NONE;
 
 /**
  * @author Dave Syer
@@ -36,7 +35,7 @@ import static org.junit.Assert.assertNotNull;
  */
 public class ConditionalOnRefreshScopeTests {
 
-	private ConfigurableApplicationContext context;
+	private AssertableApplicationContext context;
 
 	@After
 	public void close() {
@@ -47,18 +46,18 @@ public class ConditionalOnRefreshScopeTests {
 
 	@Test
 	public void refreshScopeIncluded() {
-		this.context = new SpringApplicationBuilder(RefreshAutoConfiguration.class,
-				Beans.class).web(false).bannerMode(Mode.OFF).run();
-		assertNotNull(this.context.getBean(
-				org.springframework.cloud.context.scope.refresh.RefreshScope.class));
-		assertEquals("foo", this.context.getBean("foo"));
+		this.context = AssertableApplicationContext.get(() -> new SpringApplicationBuilder(RefreshAutoConfiguration.class,
+				Beans.class).web(NONE).bannerMode(Mode.OFF).run());
+		assertThat(this.context).hasSingleBean(
+				org.springframework.cloud.context.scope.refresh.RefreshScope.class);
+		assertThat(this.context).getBean("foo").isEqualTo("foo");
 	}
 
 	@Test
 	public void refreshScopeNotIncluded() {
-		this.context = new SpringApplicationBuilder(Beans.class).web(false)
-				.bannerMode(Mode.OFF).run();
-		assertFalse(this.context.containsBean("foo"));
+		this.context = AssertableApplicationContext.get(() -> new SpringApplicationBuilder(Beans.class).web(NONE)
+				.bannerMode(Mode.OFF).run());
+		assertThat(this.context).doesNotHaveBean("foo");
 	}
 
 	@Configuration
